+++ index.html
@@ -12,9 +12,9 @@
     <link rel="shortcut icon" href="./assets/icons/favicon.svg" type="image/x-icon" id="favicon">
     <link href="https://api.fonts.coollabs.io/css2?family=Montserrat&display=swap" rel="stylesheet">
     <script type="module" src="./main.js"></script>
-    <script type="module" src="./modules/code_comparison.mjs"></script>
-    <script type="module" src="./modules/update_button.mjs"></script>
-    <script type="module" src="./modules/updated_since.mjs"></script>
+    <script type="module" src="./modules/are_codes_equal.mjs"></script>
+    <script type="module" src="./modules/quick_filter_button_change.mjs"></script>
+    <script type="module" src="./modules/change_updated_date_html.mjs"></script>
 </head>
 <body>
     <div id="menubar">
